% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Main App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input from the UI}

\item{output}{Output to the UI}

\item{session}{Session of the App}
}
\description{
Main Server of the App that contains all the modules.
}
