% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{black_theme}
\alias{black_theme}
\title{Black Theme for ggplot2}
\usage{
black_theme()
}
\description{
This function applies a custom black theme to ggplot2 plots. The theme includes a black background, white text, and other customized elements.
}
