% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{cell_type}
\alias{cell_type}
\title{Named Cell Type Vector}
\format{
An object of class \code{character} of length 1154.
}
\usage{
# Loading
data("cell_type", package = "PCAdash")
}
\description{
A named vector of Cell Type per cell Annotated using Azimuth.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
