% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{counts}
\alias{counts}
\title{Raw Counts}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 500 rows and 1154 columns.
}
\usage{
# Loading
data("counts", package = "PCAdash")
}
\description{
Raw Expression Counts in a Simple Matrix.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
