% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{edge_list_s3}
\alias{edge_list_s3}
\title{PCA Computations}
\format{
An object of class \code{list} of length 31.
}
\usage{
# Loading
data("metagene_results", package = "PCAdash")
}
\description{
S3 of PCA Results Selected Metagenes for over 80\% variance.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
