% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_xy_mod.R
\name{lt_xy_server}
\alias{lt_xy_server}
\title{Longitudinal X-Y Plot Server Module}
\usage{
lt_xy_server(
  id,
  df,
  x_col,
  y_col,
  grp_col,
  node_df,
  node_x_col,
  node_y_col,
  edge_list,
  x_label = "UMAP-1",
  y_label = "UMAP-2"
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{df}{A reactive expression that returns a data frame containing the data to be plotted.}

\item{x_col}{The column name for the x-axis.}

\item{y_col}{The column name for the y-axis.}

\item{grp_col}{The column name for the grouping variable.}

\item{node_df}{A reactive expression that returns a data frame containing the node data.}

\item{node_x_col}{The column name for the x-axis of the node data.}

\item{node_y_col}{The column name for the y-axis of the node data.}

\item{edge_list}{A reactive expression that returns a list of edge data.}

\item{x_label}{The label for the x-axis. Defaults to NULL.}

\item{y_label}{The label for the y-axis. Defaults to NULL.}
}
\description{
A Shiny server module that renders a ggplot2-based scatter plot.
}
