% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{node_df_s3}
\alias{node_df_s3}
\title{PCA Computations}
\format{
An object of class \code{data.frame} with 33 rows and 2 columns.
}
\usage{
# Loading
data("metagene_results", package = "PCAdash")
}
\description{
S3 of PCA Results Selected Metagenes for over 80\% variance.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
