% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pTime}
\alias{pTime}
\title{Named Pseudotime Vector}
\format{
An object of class \code{numeric} of length 1154.
}
\usage{
# Loading
data("pTime", package = "PCAdash")
}
\description{
A named vector of Pseudotime per cell computed using Monocle3.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
