% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tsne_coords}
\alias{tsne_coords}
\title{t-SNE Coordinates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1154 rows and 2 columns.
}
\usage{
# Loading
data("tsne_coords", package = "PCAdash")
}
\description{
t-SNE Coordinates per cell in simple Matrix.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
