% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_control_mod.R
\name{vis_params_server}
\alias{vis_params_server}
\title{Dynamic Control Server Module}
\usage{
vis_params_server(id, plot_type, max_bar_pc = reactive(5))
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}

\item{plot_type}{A reactive expression representing the selected plot type.}

\item{max_bar_pc}{A reactive expression representing the maximum number of PCs to display in the variance plot.}
}
\description{
A Shiny server module that generates reactive values based on the selected plot type.
}
