% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_control_mod.R
\name{vis_params_ui}
\alias{vis_params_ui}
\title{Dynamic control module}
\usage{
vis_params_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's}
}
\description{
A Shiny module that generates a dynamic control panel based on the selected plot type.
}
