% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{black_theme}
\alias{black_theme}
\title{Custom black theme for ggplot2 objects}
\usage{
black_theme(activate = FALSE)
}
\arguments{
\item{activate}{A logical value indicating whether to activate the theme. Defaults to FALSE.}
}
\description{
This function applies a custom black theme to ggplot2 plots.
}
