% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_xy_mod.R
\name{lt_xy_server}
\alias{lt_xy_server}
\title{Longitudinal X-Y Plot Server Module}
\usage{
lt_xy_server(
  id,
  x,
  y,
  pTime,
  main_title,
  x_label = NULL,
  y_label = NULL,
  sub_title = NULL,
  catgeory = NULL,
  cell_stroke = reactive(0.5),
  cell_size = reactive(2),
  cell_alpha = reactive(0.6),
  color_type = "contnious"
)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}

\item{x}{A numeric vector representing the x-axis data.}

\item{y}{A numeric vector representing the y-axis data.}

\item{pTime}{A numeric or datetime vector representing the progression or time data,
used for color mapping when \code{color_type} is set to "continuous".}

\item{main_title}{The main title of the plot.}

\item{x_label}{The label for the x-axis. Defaults to NULL.}

\item{y_label}{The label for the y-axis. Defaults to NULL.}

\item{sub_title}{An optional subtitle for the plot. Defaults to NULL.}

\item{catgeory}{A factor or character vector used for color mapping when \code{color_type}
is set to "discrete". Defaults to NULL.}

\item{cell_stroke}{The stroke width of the points. Defaults to 0.5.}

\item{cell_size}{The size of the points. Defaults to 2.}

\item{cell_alpha}{The transparency level of the points. Defaults to 0.6.}

\item{color_type}{A string specifying the type of color mapping. Options are "continuous"
(default) for mapping \code{pTime} to colors, or "discrete" for mapping \code{catgeory} to colors.}
}
\description{
A Shiny server module that renders a ggplot2-based scatter plot to visualize
longitudinal data over time, with options for continuous or discrete color scaling.
This module supports customization of plot aesthetics such as point size, stroke,
and transparency, along with titles and axis labels.
}
