% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts_xy_mod.R
\name{multi_ts_xy_server}
\alias{multi_ts_xy_server}
\title{X-Y Plot Server Module}
\usage{
multi_ts_xy_server(
  id,
  x,
  y_matrix,
  main_title = "Main-title",
  x_label = "x-label",
  y_label = "y-label",
  sub_title = "Sub-title",
  color_by,
  trend_width = reactive(2),
  x_breaks = reactive(5),
  cell_stroke = reactive(0.5),
  cell_size = reactive(2),
  cell_alpha = reactive(0.6),
  trend_alpha = reactive(0.8)
)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}

\item{x}{A numeric vector representing the x-axis data.}

\item{y_matrix}{A matrix of y values representing the y-axis data.}

\item{main_title}{The main title of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}

\item{sub_title}{An optional subtitle for the plot.}

\item{color_by}{A factor or character vector for coloring the points by a category.}

\item{trend_width}{The width of the trend line.}

\item{x_breaks}{The interval for breaks on the x-axis.}

\item{cell_stroke}{The stroke width of the points.}

\item{cell_size}{The size of the points.}

\item{cell_alpha}{The transparency level of the points.}

\item{trend_alpha}{The transparency level of the trend line.}
}
\description{
A Shiny server module that handles the rendering of a ggplot2-based X-Y plot.
This module accepts x and y values as inputs and provides various customization
options for rendering a detailed line plot with trend lines, titles, axis labels,
and aesthetic customizations.
}
