% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts_xy_mod.R
\name{multi_ts_xy_ui}
\alias{multi_ts_xy_ui}
\title{Multi X-Y Plot UI Module}
\usage{
multi_ts_xy_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}
}
\description{
A Shiny UI module for rendering a ggplot2-based X-Y plot. This module generates
a plot output element that can be used to display a ggplot2 plot within a Shiny application.
}
\keyword{internal}
