% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_xy_mod.R
\name{ts_xy_server}
\alias{ts_xy_server}
\title{X-Y Plot Server Module}
\usage{
ts_xy_server(
  id,
  smoothed_data,
  original_data,
  main_title = "Main Title",
  x_label = "x_lable",
  y_label = "y_lable",
  sub_title = "Sub Title"
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{smoothed_data}{A reactive expression that returns the smoothed data.}

\item{original_data}{A reactive expression that returns the original data.}

\item{main_title}{The main title of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}

\item{sub_title}{An optional subtitle for the plot.}
}
\description{
A Shiny server module that handles the rendering of a ggplot2-based X-Y plot
}
