% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_xy_mod.R
\name{ts_xy_server}
\alias{ts_xy_server}
\title{X-Y Plot Server Module}
\usage{
ts_xy_server(
  id,
  df,
  time_col,
  smoother_col,
  point_col,
  x_label = "x_lable",
  y_label = "y_lable"
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{df}{A reactive expression that returns the smoothed data.}

\item{time_col}{The column name for the time variable.}

\item{smoother_col}{The column name for the smoother variable.}

\item{point_col}{The column name for the point variable.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}
}
\description{
A Shiny server module that handles the rendering of a ggplot2-based X-Y plot
}
