% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_bar_mod.R
\name{var_bar_server}
\alias{var_bar_server}
\title{Variance Bar Plot Server Module}
\usage{
var_bar_server(
  id,
  df,
  x_col,
  y_col,
  sd_col = NULL,
  main_title,
  x_label = NULL,
  y_label = NULL,
  sub_title = NULL,
  type = "column"
)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}

\item{df}{A reactive expression that returns a data frame containing the data to be plotted.}

\item{x_col}{The column name for the x-axis.}

\item{y_col}{The column name for the y-axis.}

\item{sd_col}{The column name for the standard deviation values.}

\item{main_title}{The main title of the plot.}

\item{x_label}{The label for the x-axis. Defaults to NULL.}

\item{y_label}{The label for the y-axis. Defaults to NULL.}

\item{sub_title}{An optional subtitle for the plot. Defaults to NULL.}

\item{type}{The type of plot to render. Defaults to "column".}
}
\description{
Variance Bar Plot Server Module
}
