% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_bar_mod.R
\name{var_bar_server}
\alias{var_bar_server}
\title{Variance Bar Plot Server Module}
\usage{
var_bar_server(
  id,
  var_per_pc,
  sd,
  main_title,
  x_label = NULL,
  y_label = NULL,
  sub_title = NULL,
  bin_width = 0.5,
  bar_alpha = 0.7
)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}

\item{var_per_pc}{A numeric vector representing the variance explained by each
principal component (PC).}

\item{sd}{A numeric vector representing the standard deviation associated with
each principal component's variance.}

\item{main_title}{The main title of the plot.}

\item{x_label}{The label for the x-axis. Defaults to NULL.}

\item{y_label}{The label for the y-axis. Defaults to NULL.}

\item{sub_title}{An optional subtitle for the plot. Defaults to NULL.}

\item{bin_width}{The width of the bars in the plot. Defaults to 0.5.}

\item{bar_alpha}{The transparency level of the bars. Defaults to 0.7.}
}
\description{
A Shiny server module that renders a ggplot2-based bar plot representing the variance
explained by each principal component (PC). This module supports customization options
for titles, axis labels, and aesthetics such as bar width and transparency.
}
\details{
This module calculates the percentage of variance explained by each principal component
and generates a bar plot with error bars to indicate standard deviations. The plot is
rendered using ggplot2 with a minimalist theme and can be customized through various
parameters. It also includes options to display the percentage of variance on top of each bar.
}
