% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_bar_mod.R
\name{var_bar_ui}
\alias{var_bar_ui}
\title{Ordered Bar Plot UI Module}
\usage{
var_bar_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module, used to distinguish this module's
UI and server components from others within a Shiny application.}
}
\description{
A Shiny UI module for rendering a ggplot2-based Bar plot. This module generates
a plot output element that can be used to display a ggplot2 plot within a Shiny application.
}
